requirements.txt
google-cloud-dialogflow
pandas
numpy
scikit-learn
python-Levenshtein
openpyxl

flask==2.3.2
gunicorn==20.1.0
google-cloud-storage==2.10.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1